rename strategy package
