+++ BaseModel/src/com/wpi/cs509/teamA/strategy/AlgoStrategy.java
@@ -1,4 +1,4 @@
-package com.wpi.cs509.teamA.findRoute;
+package com.wpi.cs509.teamA.strategy;
 
 import java.util.List;
 import java.util.Map;
